Added a bunch of game related enhancements and tweaks.
Made it so costs of items are shown next to the item itself in the ingredients menu, rather than showing up when the user buys the item when prompted.
Moved the buy ingredients option from the game's main menu to the business actions menu.
Revamped the user interface in the game. The game's main menu is now grouped into several sub menus for easier navigation.