Added lots of features and tweaks.
Reorganized the Buy ingredients menu into 2 separate menus.
Made it so the menu prompts that asked weather you'd like to buy ice and salt will now show up after you type how much of a bundle you'd like to buy.
Made it so buying an ingredient bundle shows the user the cost of the bundle when hovering over the item.
Added the ability to purchase ingredient bundles to speed up the lemonade making process.