Added a bunch of new features and tweaks.
Increased The flexibility of choosing your own business titles in the game. You can now put spaces and Basic punctuation to add some depth to the title.
Added some typing sounds to all input prompts used in the game.
Fixed some small bugs with the game's auto saver.
Made it so you can no longer sell a single cup of lemonade for more than 10 dollars in cash.
Added a suddle sound that let's you know that the game has been saved.
Added the ability to autosave the game after a day has ended in the game. This also applies to setting up the game for the first time as well.