Major game enhancements.
Removed the ability to drink the lemonade that you can make and sell.
Fixed some menu navigation issues.
Added the ability to navigate to the first letter of any item in a menu.
Added the ability to search for spisific items of intrest in the new find dialog. You can press F3 in any menu to activate it.
Recoded the entire menu system for the game.