Major seller enhancements.
Made it so the number of customers that show up when selling lemonade is determined by how many cups that are poured out of the piture prior to selling it.
Revamped the lemonade seller to include a new sails popularity system. The more actual customers that show up, the higher of a chance that a single cup would be sold to them.
Modified some customer based events to be less harsh on the player.
Reduced the probability chanse of triggering random events when selling lemonade by 40 percent. It will no longer trigger every single time from now on.
Added the ability to generate random events while selling lemonade. Watch out for these, as they can be quite unexpected when they show up.